test: adding full coverage for the injector service
